refactor(webserver): extract job service
